+++ tests/Models/TypedTrackedJob.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types=1);
+
+namespace Junges\TrackableJobs\Tests\Models;
+
+use Junges\TrackableJobs\Models\TrackedJob;
+
+final class TypedTrackedJob extends TrackedJob {}